+++ src/services/Store/AddressBook/useAddressBook.tsx
@@ -1,0 +1,116 @@
+import React, { useContext, createContext, useState } from 'react';
+import isEmpty from 'lodash/isEmpty';
+
+import {
+  AddressBook,
+  ExtendedAddressBook,
+  IAccount,
+  StoreAccount,
+  LSKeys,
+  TUuid,
+  NetworkId,
+  TAddress
+} from '@types';
+import { generateUUID, isSameAddress } from '@utils';
+
+import { DataContext } from '../DataManager';
+import { useContracts } from '../Contract';
+
+export interface IAddressBookContext {
+  addressBook: ExtendedAddressBook[];
+  addressBookRestore: { [name: string]: ExtendedAddressBook | undefined };
+  createAddressBooks(addressBooksData: AddressBook): void;
+  createAddressBooksWithID(uuid: TUuid, addressBooksData: AddressBook): void;
+  updateAddressBooks(uuid: TUuid, addressBooksData: ExtendedAddressBook): void;
+  deleteAddressBooks(uuid: TUuid): void;
+  getContactByAddress(address: string): ExtendedAddressBook | undefined;
+  getContactByAddressAndNetworkId(
+    address: string,
+    networkId: NetworkId
+  ): ExtendedAddressBook | undefined;
+  getAccountLabel(account: StoreAccount | IAccount): string | undefined;
+  restoreDeletedAddressBook(addressBookId: TUuid): void;
+}
+
+function useAddressBook() {
+  const { createActions, addressBook } = useContext(DataContext);
+  const { getContractByAddress } = useContracts();
+  const [addressBookRestore, setAddressBookRestore] = useState<{
+    [name: string]: ExtendedAddressBook | undefined;
+  }>({});
+
+  const model = createActions(LSKeys.ADDRESS_BOOK);
+
+  const getContactFromContracts = (address: string): ExtendedAddressBook | undefined => {
+    const contract = getContractByAddress(address as TAddress);
+    const contact: ExtendedAddressBook | undefined = contract && {
+      address,
+      label: contract.name,
+      network: contract.networkId,
+      notes: '',
+      uuid: contract.uuid
+    };
+    return contact;
+  };
+
+  const createAddressBooks = (item: AddressBook) => model.create({ ...item, uuid: generateUUID() });
+  const createAddressBooksWithID = (uuid: TUuid, item: AddressBook) =>
+    model.create({ uuid, ...item });
+  const updateAddressBooks = (uuid: TUuid, item: ExtendedAddressBook) => model.update(uuid, item);
+  const deleteAddressBooks = (uuid: TUuid) => {
+    const addressBookToDelete = addressBook.find((a) => a.uuid === uuid);
+    if (isEmpty(addressBookToDelete)) {
+      throw new Error('Unable to delete account from address book! No account with id specified.');
+    }
+
+    setAddressBookRestore((prevState) => ({ ...prevState, [uuid]: addressBookToDelete }));
+    model.destroy(addressBookToDelete!);
+  };
+  const getContactByAddress = (address: TAddress) => {
+    return (
+      addressBook.find((contact: ExtendedAddressBook) =>
+        isSameAddress(contact.address as TAddress, address)
+      ) || getContactFromContracts(address)
+    );
+  };
+  const getContactByAddressAndNetworkId = (address: TAddress, networkId: NetworkId) => {
+    return (
+      addressBook
+        .filter((contact: ExtendedAddressBook) => contact.network === networkId)
+        .find((contact: ExtendedAddressBook) =>
+          isSameAddress(contact.address as TAddress, address)
+        ) || getContactFromContracts(address)
+    );
+  };
+  const getAccountLabel = ({ address, networkId }: { address: TAddress; networkId: NetworkId }) => {
+    const addressContact = getContactByAddressAndNetworkId(address, networkId);
+    return addressContact ? addressContact.label : undefined;
+  };
+  const restoreDeletedAddressBook = (addressBookId: TUuid) => {
+    const addressBookRecord = addressBookRestore[addressBookId];
+    if (isEmpty(addressBookRecord)) {
+      throw new Error(
+        'Unable to restore address book record! No address book record with id specified.'
+      );
+    }
+
+    const { uuid, ...rest } = addressBookRecord!;
+    createAddressBooksWithID(uuid, rest);
+    setAddressBookRestore((prevState) => ({ ...prevState, [uuid]: undefined }));
+  };
+
+  return {
+    addressBook,
+    addressBookRestore,
+    createAddressBooks,
+    createAddressBooksWithID,
+    updateAddressBooks,
+    deleteAddressBooks,
+    getContactByAddress,
+    getContactByAddressAndNetworkId,
+    getAccountLabel,
+    restoreDeletedAddressBook
+  };
+}
+
+export default useAddressBook;
